student_02_5b2ab66568be/alaska-snow-rag/config.py
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configuration
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")
PROJECT_ID = os.getenv("PROJECT_ID", "qwiklabs-gcp-02-e6e6123d96ed")

# BigQuery settings
DATASET_NAME = "alaska"
TABLE_NAME = "alaska_faq_embedded"
EMBEDDING_MODEL = "alaska.Embeddings"

# Safety settings for prompt validation
SAFETY_SETTINGS = {
    'HARM_CATEGORY_HARASSMENT': 'BLOCK_MEDIUM_AND_ABOVE',
    'HARM_CATEGORY_HATE_SPEECH': 'BLOCK_MEDIUM_AND_ABOVE',
    'HARM_CATEGORY_SEXUALLY_EXPLICIT': 'BLOCK_MEDIUM_AND_ABOVE',
    'HARM_CATEGORY_DANGEROUS_CONTENT': 'BLOCK_MEDIUM_AND_ABOVE',
}