student_02_5b2ab66568be/alaska-snow-rag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-bigquery==3.11.4
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic==2.5.0
